auth_api/python/requirements.txt
Flask
python-dotenv
pymysql
PyJWT
